#14519: Implement LlamaVision generation class which plugs into existing generation pipelines.
