+++ models/demos/llama3/demo/simple_vision_demo.py
@@ -10,8 +10,7 @@ from termcolor import cprint
 
 import llama_models.llama3.reference_impl.generation as llama_reference_generation
 from llama_models.llama3.api.tokenizer import Tokenizer
-from llama_models.llama3.api.chat_format import ChatFormat, ModelInput
-
+from llama_models.llama3.api.chat_format import ChatFormat
 from llama_models.llama3.api.datatypes import ImageMedia, UserMessage
 
 from pkg_resources import resource_filename
@@ -24,294 +23,7 @@ import os
 import ttnn
 import time
 
-
-class LlamaVision:
-    def __init__(self, model, model_args, mesh_device, vllm=False):
-        """
-        Creating a LlamaVision wrapper requires only a mesh_device and model_args.
-        With model_args you have the checkpoint location, can specify max batch size
-        and max seqlen, and other model specific parameters.
-
-        LlamaVision is general to text and chat.
-
-        For bringup, make this class general to any backend implementation, as long as it takes torch tensors and returns torch tensors.
-
-        """
-        self.model = model
-        self.model_args = model_args
-        self.mesh_device = mesh_device
-        self.vllm = vllm
-
-    def prefill_forward_single_user(
-        self,
-        vision_images,
-        vision_mask,
-        tokens,
-        xattn_caches,
-        user_id,
-        total_len,
-        prefill_len,
-    ):
-        """
-        Performs vision encode step then text prefill.
-        Returns (xattn_caches, cross_attention_masks, full_text_row_masked_out_mask, logits)
-        """
-        B = tokens.shape[0]
-        xattn_caches, cross_attention_masks, full_text_row_masked_out_mask = self.model.compute_vision_tokens_masks(
-            batch_images=[vision_images],
-            batch_masks=[vision_mask],
-            total_len=total_len,
-            xattn_caches=xattn_caches,
-            user_id=user_id,
-        )
-
-        (
-            tt_h,
-            tt_xattn_mask,
-            tt_full_text_mask_expand_1NSH,
-            tt_full_text_mask_expand_11SD,
-            tt_position_id,
-            rot_mats,
-            transformation_mats,
-        ) = self.model.prepare_inputs_prefill(
-            tokens, cross_attention_masks, full_text_row_masked_out_mask, prefill_len=prefill_len
-        )
-
-        tt_logits = self.model.ttnn_prefill_forward(
-            tt_h,
-            tt_xattn_mask,
-            tt_full_text_mask_expand_1NSH,
-            tt_full_text_mask_expand_11SD,
-            xattn_caches,
-            tt_position_id,
-            rot_mats,
-            transformation_mats,
-            user_id,
-        )
-
-        logits = self.model.process_output_prefill(tt_logits, B, prefill_len)
-
-        return xattn_caches, cross_attention_masks, full_text_row_masked_out_mask, logits
-
-    def decode_forward(
-        self,
-        position_id,
-        tokens,
-        cross_attention_masks,
-        full_text_row_masked_out_mask,
-        xattn_caches,
-    ):
-        """
-        Performs text decode step.
-        Returns logits
-        """
-
-        # forward_decode should be traced callable
-        # decorator does compilation, capture, execute
-        # B = 1 # TODO: Only supports batch=1 right now! Might make tokens input a tensor.
-        # S = 1
-        B, S = tokens.shape
-
-        (
-            tt_h,
-            tt_xattn_mask,
-            tt_full_text_mask_expand_1NSH,
-            _,
-            tt_position_id,
-            rot_mats,
-            _,
-        ) = self.model.prepare_inputs_decode(
-            tokens, cross_attention_masks, full_text_row_masked_out_mask, position_id=position_id
-        )
-
-        tt_logits = self.model.ttnn_decode_forward(
-            tt_h,
-            tt_xattn_mask,
-            tt_full_text_mask_expand_1NSH,
-            xattn_caches,
-            tt_position_id,
-            rot_mats,
-        )
-
-        logits = self.model.process_output_decode(tt_logits, B, S)
-        return logits
-
-    def capture_trace(
-        self,
-        position_id,
-        tokens,
-        cross_attention_masks,
-        full_text_row_masked_out_mask,
-        xattn_caches,
-    ):
-        """
-        Captures a trace for the decode_forward method.
-        """
-        (
-            tt_h,
-            tt_xattn_mask,
-            tt_full_text_mask_expand_1NSH,
-            _,
-            tt_position_id,
-            rot_mats,
-            _,
-        ) = self.model.prepare_inputs_decode(
-            tokens, cross_attention_masks, full_text_row_masked_out_mask, position_id=position_id
-        )
-
-        # Compile run
-        tt_logits_rm = self.model.ttnn_decode_forward(
-            tt_h,
-            tt_xattn_mask,
-            tt_full_text_mask_expand_1NSH,
-            xattn_caches,
-            tt_position_id,
-            rot_mats,
-        )
-
-        # Get inputs ready for trace run
-        (
-            tt_h,
-            tt_xattn_mask,
-            tt_full_text_mask_expand_1NSH,
-            _,
-            tt_position_id,
-            rot_mats,
-            _,
-        ) = self.model.prepare_decode_inputs_host(
-            tokens, cross_attention_masks, full_text_row_masked_out_mask, position_id
-        )
-
-        (
-            tt_h,
-            tt_xattn_mask,
-            tt_full_text_mask_expand_1NSH,
-            tt_position_id,
-            rot_mats,
-        ) = self.model.copy_host_to_device(
-            (tt_h, tt_xattn_mask, tt_full_text_mask_expand_1NSH, tt_position_id, rot_mats)
-        )
-
-        trace_id = ttnn.begin_trace_capture(self.mesh_device, cq_id=0)
-        B = tokens.shape[0]
-        # Do on-device transformations of inputs before forward
-        tt_xattn_mask_transform, tt_full_text_mask_expand_1NSH_transform = self.model.transform_decode_inputs_device(
-            tt_xattn_mask,
-            tt_full_text_mask_expand_1NSH,
-            B=B,
-        )
-
-        tt_logits_rm = self.model.ttnn_decode_forward(
-            tt_h,
-            tt_xattn_mask_transform,
-            tt_full_text_mask_expand_1NSH_transform,
-            xattn_caches,
-            tt_position_id,
-            rot_mats,
-        )
-
-        ttnn.end_trace_capture(self.mesh_device, trace_id, cq_id=0)
-
-        return trace_id, tt_logits_rm, tt_h, tt_xattn_mask, tt_full_text_mask_expand_1NSH, tt_position_id, rot_mats
-
-    def decode_forward_trace(
-        self,
-        position_id,
-        tokens,
-        cross_attention_masks,
-        full_text_row_masked_out_mask,
-        xattn_caches,  # TODO: unused since captured in trace?
-        trace_id,
-        trace_logits_rm,
-        trace_h,
-        trace_xattn_mask,
-        trace_full_text_mask_expand_1NSH,
-        trace_position_id,
-        trace_rot_mats,
-    ):
-        (
-            tt_h,
-            tt_xattn_mask,
-            tt_full_text_mask_expand_1NSH,
-            _,
-            tt_position_id,
-            rot_mats,
-            _,
-        ) = self.model.prepare_decode_inputs_host(
-            tokens, cross_attention_masks, full_text_row_masked_out_mask, position_id=position_id
-        )
-
-        self.model.copy_host_to_device(
-            host_tensors=(tt_h, tt_xattn_mask, tt_full_text_mask_expand_1NSH, tt_position_id, rot_mats),
-            device_tensors=(
-                trace_h,
-                trace_xattn_mask,
-                trace_full_text_mask_expand_1NSH,
-                trace_position_id,
-                trace_rot_mats,
-            ),
-        )
-
-        ttnn.execute_trace(self.mesh_device, trace_id, cq_id=0, blocking=False)
-
-        B, S = tokens.shape
-        logits = self.model.process_output_decode(trace_logits_rm, B=B, S=S)
-
-        return logits
-
-    def easy_trace(
-        self,
-        position_id,
-        tokens,
-        cross_attention_masks,
-        full_text_row_masked_out_mask,
-        xattn_caches,
-    ):
-        """
-        Tracing is easy! Just call this method and you'll run traced
-        """
-        if not hasattr(self, "trace_id"):
-            (
-                trace_id,
-                tt_logits_rm,
-                tt_h,
-                tt_xattn_mask,
-                tt_full_text_mask_expand_1NSH,
-                tt_position_id,
-                rot_mats,
-            ) = self.capture_trace(
-                position_id,
-                tokens,
-                cross_attention_masks,
-                full_text_row_masked_out_mask,
-                xattn_caches,
-            )
-            self.trace_id = trace_id
-            self.trace_inputs = {
-                "tt_h": tt_h,
-                "tt_xattn_mask": tt_xattn_mask,
-                "tt_full_text_mask_expand_1NSH": tt_full_text_mask_expand_1NSH,
-                "tt_position_id": tt_position_id,
-                "rot_mats": rot_mats,
-            }
-            self.trace_outputs = {
-                "tt_logits_rm": tt_logits_rm,
-            }
-
-        return self.decode_forward_trace(
-            position_id,
-            tokens,
-            cross_attention_masks,
-            full_text_row_masked_out_mask,
-            xattn_caches,
-            self.trace_id,
-            self.trace_outputs["tt_logits_rm"],
-            self.trace_inputs["tt_h"],
-            self.trace_inputs["tt_xattn_mask"],
-            self.trace_inputs["tt_full_text_mask_expand_1NSH"],
-            self.trace_inputs["tt_position_id"],
-            self.trace_inputs["rot_mats"],
-        )
+from models.demos.llama3.tt.multimodal.vision_generator import LlamaVision
 
 
 def get_sampler(temperature, top_p, tokenizer):
@@ -370,11 +82,17 @@ def create_multimodal_model(mesh_device, max_batch_size, max_seq_len, dtype=ttnn
         "normal",
     ],
 )
+@pytest.mark.parametrize(
+    "enable_trace",
+    (False, True),
+    ids=["no_trace", "trace"],
+)
 @pytest.mark.parametrize("device_params", [{"trace_region_size": 14951424, "num_command_queues": 2}], indirect=True)
 def test_llama_multimodal_demo_text(
     mesh_device,
     warmup_iters,
     test_case,
+    enable_trace,
     temperature: float = 0,
     top_p: float = 0.9,
     max_seq_len: int = 512,
@@ -391,11 +109,11 @@ def test_llama_multimodal_demo_text(
     mesh_device.enable_program_cache()
     mesh_device.enable_async(True)
     model_args, model = create_multimodal_model(mesh_device, max_batch_size=max_batch_size, max_seq_len=max_seq_len)
-    model = LlamaVision(model, model_args, mesh_device)
+    generator = LlamaVision(model, model_args, mesh_device)
     tokenizer = Tokenizer(model_path=tokenizer_path)
     formatter = ChatFormat(tokenizer)
 
-    xattn_caches = model.model.setup_cache(model_args.max_batch_size)
+    xattn_caches = generator.model.setup_cache(model_args.max_batch_size)
 
     with open(IMG_PATH / "dog.jpg", "rb") as f:
         img = PIL_Image.open(f).convert("RGB")
@@ -447,7 +165,7 @@ def test_llama_multimodal_demo_text(
                 cross_attention_masks,
                 full_text_row_masked_out_mask,
                 logits,
-            ) = model.prefill_forward_single_user(
+            ) = generator.prefill_forward_single_user(
                 vision_images,
                 vision_mask,
                 prompt_tokens_tensor,
@@ -459,57 +177,35 @@ def test_llama_multimodal_demo_text(
             prefill_end = time.perf_counter()
 
             next_token, text = sampler(logits)
-            # logger.info(f"Prefill output: {next_token}:{text}")
             tokens[0, prefill_len] = next_token
 
             decode_times = []
-
-            # Capture trace
-            # next_token_tensor = torch.tensor([next_token], dtype=torch.long).reshape(1, 1)  # B, S
-            # trace_id, tt_logits_rm, tt_h, tt_xattn_mask, tt_full_text_mask_expand_1NSH, tt_position_id, rot_mats = model.capture_trace(
-            #     prefill_len,
-            #     next_token_tensor,
-            #     cross_attention_masks,
-            #     full_text_row_masked_out_mask,
-            #     xattn_caches,
-            # )
 
             for gen_idx in range(max_gen_len - 1):
                 decode_start = time.perf_counter()
                 position_id = prefill_len + gen_idx
                 next_token_tensor = torch.tensor([next_token], dtype=torch.long).reshape(1, 1)  # B, S
-                # logits = model.decode_forward(
-                #     position_id,
-                #     next_token_tensor,
-                #     cross_attention_masks,
-                #     full_text_row_masked_out_mask,
-                #     xattn_caches,
-                # )
-                logits = model.easy_trace(
-                    position_id,
-                    next_token_tensor,
-                    cross_attention_masks,
-                    full_text_row_masked_out_mask,
-                    xattn_caches,
-                )
-                # logits = model.decode_forward_trace(
-                #     position_id,
-                #     next_token_tensor,
-                #     cross_attention_masks,
-                #     full_text_row_masked_out_mask,
-                #     xattn_caches,
-                #     trace_id,
-                #     tt_logits_rm,
-                #     tt_h,
-                #     tt_xattn_mask,
-                #     tt_full_text_mask_expand_1NSH,
-                #     tt_position_id,
-                #     rot_mats
-                # )
+
+                if enable_trace:
+                    logits = generator.easy_trace(
+                        position_id,
+                        next_token_tensor,
+                        cross_attention_masks,
+                        full_text_row_masked_out_mask,
+                        xattn_caches,
+                    )
+                else:
+                    logits = generator.decode_forward(
+                        position_id,
+                        next_token_tensor,
+                        cross_attention_masks,
+                        full_text_row_masked_out_mask,
+                        xattn_caches,
+                    )
+
                 next_token, text = sampler(logits)
                 # Update next token
                 tokens[0, position_id + 1] = next_token
-                # logger.info(f"Decode output {position_id}: {next_token}:{text}")
                 decode_end = time.perf_counter()
                 decode_times.append(decode_end - decode_start)
 
@@ -530,4 +226,4 @@ def test_llama_multimodal_demo_text(
             decode_time_ms = sum(decode_times) / (gen_idx + 1) * 1000
             logger.info(f"Decode time: {decode_time_ms:.2f} ms")
 
-            # ttnn.release_trace(model.mesh_device, trace_id)
+            # ttnn.release_trace(generator.mesh_device, trace_id)
